Change int to double in input StandarDeviasiSample.java

+++ StandarDeviasiSample/StandarDeviasiSample.java
@@ -15,25 +15,25 @@ public class StandarDeviasiSample {
         int size = scan.nextInt();
 
         // Buat array dengan size yang diinput
-        int dataArray[] = new int[size];
+        double dataArray[] = new double[size];
 
         // Isikan dataArray
         for(int i = 0; i < size; i++) {
             System.out.print("Masukkan data ke-" + (i + 1) + ": ");
-            dataArray[i] = scan.nextInt();
+            dataArray[i] = scan.nextDouble();
         }
 
         // Tutup scanner
         scan.close();
 
         // Jumlahkan isi dataArray
-        int sum = 0;
+        double sum = 0;
         for(int i = 0; i < size; i++) {
             sum += dataArray[i];
         }
 
         // Rata-rata array
-        double average = (double)sum / (double)size;
+        double average = sum / size;
 
 
         // Perhitungan pembilang
